requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.17.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
courlan==1.0.0
cssselect==1.2.0
dataclasses-json==0.6.4
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
elevenlabs==1.0.0b1
executing==2.0.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
htmldate==1.7.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.21.3
hyperframe==6.0.1
idna==3.6
ipython==8.22.2
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jusText==3.0.0
langcodes==3.3.0
llama-index-core==0.10.16.post1
llama-index-embeddings-huggingface==0.1.4
llama-index-embeddings-openai==0.1.6
llama-index-llms-anthropic==0.1.5
llama-index-llms-mistralai==0.1.4
llama-index-llms-openai==0.1.7
llama-index-readers-file==0.1.7
llama-index-readers-json==0.1.4
llama-index-readers-web==0.1.6
llama-index-vector-stores-qdrant==0.1.3
llamaindex-py-client==0.1.13
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistralai==0.0.11
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.13.3
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
playwright==1.42.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
pyee==11.0.1
Pygments==2.17.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.2
pypdf==4.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.7.3
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
selenium==4.18.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
streamlit==1.32.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tinysegmenter==0.3
tld==0.13
tldextract==5.1.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.38.2
tree-sitter==0.20.4
tree-sitter-languages==1.10.2
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
